binBoundaries ArrayList<Number> ivm state uit GWT/HTML5, ihkv DWOWIDGET-65
git-svn-id: https://svn.science.uu.nl/repos/project.fisme.java/Statistiek/trunk@15452 58dd2db4-0f11-0410-9df4-da15f00f50bd

+++ src/fi/statistiek/descriptives/DescriptivesUserOptionsPanel.java
@@ -470,15 +470,15 @@ public class DescriptivesUserOptionsPanel extends JPanel implements
     				|| splitType.equals(AllowedTypes.INTEGER))
     			{
     				this.splitMinBoundaryField.setText(
-    					Statistiek.getStringValue(this.model.getSplitOptions().getBinBoundaries().get(0)));
+    					Statistiek.getStringValue(this.model.getSplitOptions().getBinBoundaries().get(0).doubleValue()));
     				this.splitBinWidthField.setText(Statistiek.getFormattedBinWidth(this.model.getSplitOptions().getBinBoundaries()));
     				StringBuilder sb = new StringBuilder();
     				for (int i = 0; i < this.model.getSplitOptions()
     					.getBinBoundaries().size() - 1; i++)
     				{
-    					sb.append(Statistiek.getStringValue(this.model.getSplitOptions().getBinBoundaries().get(i)));
+    					sb.append(Statistiek.getStringValue(this.model.getSplitOptions().getBinBoundaries().get(i).doubleValue()));
     					sb.append(" -< ");
-    					sb.append(Statistiek.getStringValue(this.model.getSplitOptions().getBinBoundaries().get(i + 1)));
+    					sb.append(Statistiek.getStringValue(this.model.getSplitOptions().getBinBoundaries().get(i + 1).doubleValue()));
     					sb.append("\n");
     				}
     				this.splitBoundariesArea.setText(sb.toString());
